requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
ldap3==2.9.1
Mako==1.3.6
MarkupSafe==3.0.2
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
